enable preserve-argv0 by default on mainline
P flag is needed in order for qemu to not lose
argv0 value. Problem is that qemu can only detect that
flag was set on 5.12+ kernels and if it doesn’t detect
it then all argument values are wrong.

This introduces a patch that is enabled on mainline
target that assumes P flag was set if the kernel does
not provide any flags.

Signed-off-by: Tonis Tiigi <tonistiigi@gmail.com>