patches: Avoid executable path resolution
The buildkit inline emulator resolves paths of the binary it executes.
This is needed for readlink()/open()/... to work on /proc/self/exe
correctly, in case the binary changes its working directory. However the
resolution may result in unexpected args values. So avoid passing the
resolved paths around, but use them when the emulated process accesses
/proc/self/exe.

Known issue - a shell script like `#!./chwd` will still not work as expected,
as it will set /proc/self/exe to the script, rather than the actual
executable.

Signed-off-by: Piotr Stankiewicz <piotr.stankiewicz@docker.com>